// projects/ui/src/lib/components/po-info/po-info-orietation.enum.ts
/**
 * @usedBy PoInfoComponent
 *
 * @description
 *
 * Define os tipos de orientações disponíveis para o `po-info`.
 */
export enum PoInfoOrientation {
  /** O valor será exibido na horizontal, ao lado direito em relação ao label. */
  Horizontal = 'horizontal',

  /** Exibe o valor na vertical, ou seja, abaixo do label. */
  Vertical = 'vertical'
}